# OpenApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "digitalekrankheitsanamnese _31_b_ID1",
    "version": "V1"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Returns the list of registered users",
        "operationId": "getApiV1Users",
        "responses": {
          "200": {
            "description": "successful response",
            "content": {
              "application/vnd.de.health.users+json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                },
                "examples": {
                  "List of registered users": {
                    "value": {"users":[{"userId":"0","name":"Max","surname":"Mustermann","socialId":"DE12345678","self":{"href":"/api/v1/users/0"},"questionnaires":{"href":"/api/v1/users/0/questionnaires"},"all":{"href":"/api/v1/users"},"update_user":{"href":"/api/v1/users/0","action":"PUT"}}],"self":{"href":"/api/v1/users"},"register_user":{"href":"/api/v1/users","action":"POST"}}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Registers a new user",
        "operationId": "postApiV1Users",
        "requestBody": {
          "description": "Name, surname and social ID",
          "content": {
            "application/vnd.de.health.user+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "Data": {
                  "value": {"name":"Max","surname":"Musternmann","socialId":"DE12345678"}
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "headers": {
              "Location": {
                "description": "Contains the URL pointing to the resource representing the user that has been successfully registered",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Returns the user with the specified userId",
        "operationId": "getApiV1UsersByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The ID of the user",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull response",
            "content": {
              "application/vnd.de.health.user+json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "examples": {
                  "User": {
                    "value": {"userId":"0","name":"Max","surname":"Mustermann","socialId":"DE12345678","self":{"href":"/api/v1/users/0"},"questionnaires":{"href":"/api/v1/users/0/questionnaires"},"all":{"href":"/api/v1/users"},"update_user":{"href":"/api/v1/users/0","action":"PUT"}}
                  }
                }
              }
            }
          },
          "404": {
            "description": "The user with the passed ID does not exist",
            "content": {
              "application/vnd.de.health.error+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "User not found": {
                    "value": {"code":404,"message":"The user having the 'userId':'{VALUE}' does not exist"}
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Updates the user profile with the specified userId",
        "operationId": "putApiV1UsersByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "The ID of the user",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Name, surname and social ID",
          "content": {
            "application/vnd.de.health.user+json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "examples": {
                "Data": {
                  "value": {"name":"Max","surname":"Musternmann","socialId":"DE12345678"}
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The response contains the updated user",
            "content": {
              "application/vnd.de.health.user+json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "examples": {
                  "User": {
                    "value": {"userId":"0","name":"Max","surname":"Mustermann","socialId":"DE12345678","self":{"href":"/api/v1/users/0"},"questionnaires":{"href":"/api/v1/users/0/questionnaires"},"all":{"href":"/api/v1/users"},"update_user":{"href":"/api/v1/users/0","action":"PUT"}}
                  }
                }
              }
            }
          },
          "404": {
            "description": "The user with the passed ID does not exist",
            "content": {
              "application/vnd.de.health.error+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                },
                "examples": {
                  "User not found": {
                    "value": {"code":404,"message":"The user having the 'userId':'{VALUE}' does not exist"}
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Hyperlink": {
        "type": "object",
        "properties": {
          "rel": {
            "type": "string",
            "description": "The IANA relation type (optional)"
          },
          "desc": {
            "type": "string",
            "description": "A description providing further details about the semantic relation (optional)"
          },
          "href": {
            "type": "string",
            "description": "The URI addressing the related resource"
          },
          "templated": {
            "type": "boolean",
            "description": "A flag indicating whether 'href' contains a URI-Template describing a range of URIs (see RFC6570)",
            "nullable": true
          },
          "action": {
            "type": "string",
            "description": "The action (i.e. HTTP method) a client should apply when following this hyperlink"
          }
        }
      },
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "The HTTP status code (the three digit code) of the error message",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "The message providing further details concerning the error"
          }
        },
        "example": {"code":404,"message":"The user having the 'userId':'{VALUE}' does not exist"}
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "The ID of the User"
          },
          "name": {
            "type": "string",
            "description": "The name of the User"
          },
          "surname": {
            "type": "string",
            "description": "The surname of the User"
          },
          "socialId": {
            "type": "string",
            "description": "The social ID of the User"
          },
          "self": {
            "$ref": "#/components/schemas/Hyperlink"
          },
          "questionnaires": {
            "$ref": "#/components/schemas/Hyperlink"
          },
          "all": {
            "$ref": "#/components/schemas/Hyperlink"
          },
          "update_user": {
            "$ref": "#/components/schemas/Hyperlink"
          }
        },
        "example": {"userId":"0","name":"Max","surname":"Mustermann","socialId":"DE12345678","self":{"href":"/api/v1/users/0"},"questionnaires":{"href":"/api/v1/users/0/questionnaires"},"all":{"href":"/api/v1/users"},"update_user":{"href":"/api/v1/users/0","action":"PUT"}}
      },
...
      "Users": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "description": "A collection of Users"
          },
          "self": {
            "$ref": "#/components/schemas/Hyperlink"
          },
          "register_user": {
            "$ref": "#/components/schemas/Hyperlink"
          }
        },
        "example": {"users":[{"userId":"0","name":"Max","surname":"Mustermann","socialId":"DE12345678","self":{"href":"/api/v1/users/0"},"questionnaires":{"href":"/api/v1/users/0/questionnaires"},"all":{"href":"/api/v1/users"},"update_user":{"href":"/api/v1/users/0","action":"PUT"}}],"self":{"href":"/api/v1/users"},"register_user":{"href":"/api/v1/users","action":"POST"}}
      }
    }
  }
}